test_utils: extract make_throwing_stream as a test_util
(cherry picked from commit be77f709bce76f2af759d7d3c3259f6af1aeecde)
