+++ src/v/utils/tests/input_stream_fanout_test.cc
@@ -13,6 +13,7 @@
 #include "bytes/iostream.h"
 #include "bytes/random.h"
 #include "random/generators.h"
+#include "test_utils/iostream.h"
 #include "utils/stream_utils.h"
 
 #include <seastar/core/abort_source.hh>
@@ -437,29 +438,8 @@ SEASTAR_THREAD_TEST_CASE(input_stream_fanout_detach_10_size_limit) {
     test_detached_consumer<10>(4, 1000);
 }
 
-template<class Err>
-ss::input_stream<char> make_throwing_stream(Err err) {
-    struct throwing_stream final : ss::data_source_impl {
-        explicit throwing_stream(Err e)
-          : _err(std::move(e)) {}
-
-        ss::future<ss::temporary_buffer<char>> skip(uint64_t) final {
-            return get();
-        }
-
-        ss::future<ss::temporary_buffer<char>> get() final {
-            return ss::make_exception_future<ss::temporary_buffer<char>>(
-              std::move(_err));
-        }
-
-        Err _err;
-    };
-    auto ds = ss::data_source(std::make_unique<throwing_stream>(err));
-    return ss::input_stream<char>(std::move(ds));
-}
-
 SEASTAR_THREAD_TEST_CASE(input_stream_fanout_producer_throw) {
-    auto is = make_throwing_stream(ss::abort_requested_exception());
+    auto is = tests::make_throwing_stream(ss::abort_requested_exception());
     auto [s1, s2] = input_stream_fanout<2>(std::move(is), 4, 8);
 
     BOOST_REQUIRE_THROW(s1.read().get(), ss::abort_requested_exception);
